test: isolate construction checker coverage
